Move timer handler in timer module mod.rs
