+++ xernel/kernel/src/timer/timer_queue.rs
@@ -1,20 +1,14 @@
 use core::time::Duration;
-
+use crate::timer::TIMER_VECTOR;
 use crate::arch::amd64::apic::APIC;
-use crate::arch::amd64::interrupts::register_handler;
-use crate::arch::amd64::tsc;
-use crate::cpu::current_cpu;
-use crate::sched::context::TrapFrame;
 use crate::timer::timer_event::EventExecutor;
 use crate::timer::timer_event::TimerEvent;
 use alloc::collections::VecDeque;
 use alloc::vec::Vec;
-use libxernel::sync::Once;
 
-static TIMER_VECTOR: Once<u8> = Once::new();
 
 pub struct TimerQueue {
-    events: VecDeque<TimerEvent>,
+    pub events: VecDeque<TimerEvent>,
     pub is_timer_set: bool,
 }
 
@@ -89,37 +83,3 @@ impl TimerQueue {
     }
 }
 
-pub fn init() {
-    tsc::calibrate_tsc();
-
-    TIMER_VECTOR.set_once(0xE0);
-
-    register_handler(*TIMER_VECTOR, timer_interrupt_handler);
-}
-
-pub fn timer_interrupt_handler(_frame: &mut TrapFrame) {
-    // if periodic, add again to queue
-    // set timer to next event in queue
-
-    let cpu = current_cpu();
-
-    let mut timer_queue = cpu.timer_queue.write();
-
-    //timer_queue.deadlines();
-
-    timer_queue.event_dispatch();
-
-    let next_event = timer_queue.events.front();
-
-    if let Some(event) = next_event {
-        APIC.oneshot(*TIMER_VECTOR, &event.deadline);
-
-        if event.periodic {
-            //timer_queue.queue_event(event.clone());
-        }
-    } else {
-        // No event in event queue?
-    }
-
-    timer_queue.unlock();
-}
